requirements.txt
influxdb
pyfunctional